#30: add: router from creating room to game room
